rename debug, split off center_capture_hole
